code cleanup

+++ SEDMr/Cube.py
@@ -265,8 +265,8 @@ def extraction_to_cube(exts, outname="G.npy"):
     xreflams = reject_outliers(
             np.array([ex.xreflam for ex in exts], dtype=np.float))
     meta_data = {"fiducial_wavelength": np.mean(xreflams)}
-    print("Avg lam, Std lam: %f, %f" % (np.mean(xreflams),
-                                        np.std(xreflams)))
+    print("Avg lam, Std lam: %f, %f" % (float(np.mean(xreflams)),
+                                        float(np.std(xreflams))))
 
     # make arrays
     xs = np.array(xs, dtype=np.float)
